Change logger abstraction to macros due to constant expression restriction of string format

+++ skymarlin/utility/MutableByteBuffer.test.cpp
@@ -31,7 +31,7 @@ TEST(MutableByteBuffer, InvalidStringSize)
 
     try {
         std::string s {};
-        for (int i = 0; i < buffer_size + 10; ++i) {
+        for (size_t i = 0; i < buffer_size + 10; ++i) {
             s += "k";
         }
